+++ guillotina/catalog/utils.py
@@ -1,4 +1,4 @@
-from guillotina.content import iter_schemata_for_type_extended
+from guillotina.content import get_all_possible_schemas_for_type
 from guillotina.directives import index
 from guillotina.directives import merged_tagged_value_dict
 from guillotina.directives import merged_tagged_value_list
@@ -7,7 +7,7 @@ from guillotina.directives import metadata
 
 def get_index_fields(type_name):
     mapping = {}
-    for schema in iter_schemata_for_type_extended(type_name):
+    for schema in get_all_possible_schemas_for_type(type_name):
         # create mapping for content type
         mapping.update(merged_tagged_value_dict(schema, index.key))
     return mapping
@@ -15,7 +15,7 @@ def get_index_fields(type_name):
 
 def get_metadata_fields(type_name):
     fields = []
-    for schema in iter_schemata_for_type_extended(type_name):
+    for schema in get_all_possible_schemas_for_type(type_name):
         # create mapping for content type
         fields.extend(merged_tagged_value_list(schema, metadata.key))
     return fields